services/todo-service/requirements.txt
# services/todo-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
requests==2.31.0
email-validator==2.1.0